requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-redis-cache==1.5.5
django-tastypie==0.12.2
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
redis==2.10.3
six==1.9.0
static3==0.6.1
trueskill==0.4.3
wheel==0.24.0
whitenoise==2.0.2
